Move the current application to /classic-2024 for further development
