--- src/app/theme.tsx
@@ -1,6 +1,0 @@
-// src/theme.ts
-import { createTheme } from "@mantine/core";
-
-export const theme = createTheme({
-  // ... other theme override properties
-});
