--- src/components/RequirementTable/RequirementTable.module.css
@@ -1,10 +1,0 @@
-.card {
-  background-color: light-dark(
-    var(--mantine-color-gray-0),
-    var(--mantine-color-dark-6)
-  );
-}
-.title {
-  font-family: Greycliff CF, var(--mantine-font-family);
-  font-weight: 700;
-}
